# packages/specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.15.0
  description: |
    Spec-first HTTP API. Paths include the versioned prefix (`/api/v1`) so
    generated clients can be used without additional base-URL rewriting.

    **Authentication endpoints** live under `/api/v1/auth/*` and follow the
    [Better Auth protocol](https://www.better-auth.com/docs). They are
    deliberately not enumerated here: Better Auth owns that wire contract and
    evolves it independently of our OpenAPI surface. The `cookieAuth` and
    `bearerAuth` security schemes documented below describe how the *rest* of
    this API accepts credentials issued by that namespace — session cookies
    (web) and bearer tokens (mobile, via the `bearer` plugin).
  contact:
    name: App Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder)
...
paths:
...
  /api/v1/admin/instructors:
    post:
      operationId: upsertInstructor
      summary: Create or update an instructor
      description: |
        Creates a new instructor record or updates an existing one (matched by
        slug or an externalId collision). Returns 409 when the provided slug
        already exists and belongs to a *different* instructor than would be
        matched by externalIds. Requires admin role.
      tags: [admin, Entities]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertInstructorRequest"
            examples:
              create:
                summary: Create Andrei Neagoie with a Udemy externalId
                value:
                  displayName: Andrei Neagoie
                  slug: andrei-neagoie
                  externalIds:
                    - source: udemy
                      externalId: "udemy:instructor:42"
                      url: "https://udemy.com/user/42"
      responses:
        "201":
          description: Instructor created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorDto"
              examples:
                created:
                  value:
                    id: clxvins0000000000000000001
                    slug: andrei-neagoie
                    displayName: Andrei Neagoie
                    externalIds:
                      - source: udemy
                        externalId: "udemy:instructor:42"
                        url: "https://udemy.com/user/42"
                    coursesTotal: 0
                    createdAt: "2026-05-23T10:00:00Z"
                    updatedAt: "2026-05-23T10:00:00Z"
        "400":
          description: Validation error — missing or malformed fields
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badInput:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "displayName must not be empty."
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have the admin role
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: Administrator role required.
        "409":
          description: Slug already taken by a different instructor
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                conflict:
                  value:
                    type: about:blank
                    title: Conflict
                    status: 409
                    detail: "Slug 'andrei-neagoie' is already taken by a different instructor."
  /api/v1/admin/maintenance/backfill-metadata:
    post:
      operationId: startBackfillMetadata
      summary: Trigger a background metadata backfill across the library
      description: |
        Enqueues a background job that walks every course in the specified
        library (or all libraries when `libraryId` is omitted), reads each
        course's `course.json`, and upserts instructor/studio/tag links and
        extended fields. Returns 202 immediately with a `jobId`; subscribe to
        the `maintenance:backfill:{jobId}` Centrifugo channel to track progress.
        Admin only.
      tags: [admin, Maintenance]
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackfillMetadataRequest"
            examples:
              allLibraries:
                summary: Backfill all libraries
                value: {}
              oneLibrary:
                summary: Backfill a specific library
                value:
                  libraryId: clxvlib0000000000000000001
      responses:
        "202":
          description: Backfill job accepted — progress via Centrifugo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackfillJobAccepted"
              examples:
                accepted:
                  value:
                    jobId: clxvjob0000000000000000001
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have the admin role
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: Administrator role required.
  /api/v1/admin/studios:
    post:
      operationId: upsertStudio
      summary: Create or update a studio
      description: |
        Creates a new studio record or updates an existing one (matched by slug
        or externalId collision). Returns 409 when the slug is taken by a
        different studio. Requires admin role.
      tags: [admin, Entities]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertStudioRequest"
            examples:
              create:
                summary: Create Zero To Mastery studio
                value:
                  displayName: Zero To Mastery
                  slug: zero-to-mastery
                  externalIds:
                    - source: udemy
                      externalId: "udemy:organization:ztm"
                      url: "https://udemy.com/organization/ztm"
      responses:
        "201":
          description: Studio created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioDto"
              examples:
                created:
                  value:
                    id: clxvstu0000000000000000001
                    slug: zero-to-mastery
                    displayName: Zero To Mastery
                    externalIds:
                      - source: udemy
                        externalId: "udemy:organization:ztm"
                        url: "https://udemy.com/organization/ztm"
                    coursesTotal: 0
                    createdAt: "2026-05-23T10:00:00Z"
                    updatedAt: "2026-05-23T10:00:00Z"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badInput:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "displayName must not be empty."
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have the admin role
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: Administrator role required.
        "409":
          description: Slug already taken by a different studio
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                conflict:
                  value:
                    type: about:blank
                    title: Conflict
                    status: 409
                    detail: "Slug 'zero-to-mastery' is already taken by a different studio."
  /api/v1/admin/tags:
    post:
      operationId: upsertTag
      summary: Create or update a tag
      description: |
        Creates a new tag or updates an existing one (matched by slug or
        externalId collision). Returns 409 when the slug is taken by a
        different tag. Requires admin role.
      tags: [admin, Entities]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTagRequest"
            examples:
              create:
                summary: Create a JavaScript tag in the language category
                value:
                  displayName: JavaScript
                  slug: javascript
                  category: language
      responses:
        "201":
          description: Tag created or updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
              examples:
                created:
                  value:
                    id: clxvtag0000000000000000001
                    slug: javascript
                    displayName: JavaScript
                    category: language
                    externalIds: []
                    coursesTotal: 0
                    createdAt: "2026-05-23T10:00:00Z"
                    updatedAt: "2026-05-23T10:00:00Z"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                badInput:
                  value:
                    type: about:blank
                    title: Bad Request
                    status: 400
                    detail: "displayName must not be empty."
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
        "403":
          description: Caller does not have the admin role
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                forbidden:
                  value:
                    type: about:blank
                    title: Forbidden
                    status: 403
                    detail: Administrator role required.
        "409":
          description: Slug already taken by a different tag
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                conflict:
                  value:
                    type: about:blank
                    title: Conflict
                    status: 409
                    detail: "Slug 'javascript' is already taken by a different tag."
...
  /api/v1/catalog/instructors:
    get:
      operationId: listInstructors
      summary: List instructors with optional search
      description: |
        Returns a paginated list of instructors. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. Results are
        ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
      responses:
        "200":
          description: Instructor list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvins0000000000000000001
                        slug: andrei-neagoie
                        displayName: Andrei Neagoie
                        externalIds:
                          - source: udemy
                            externalId: "udemy:instructor:42"
                            url: "https://udemy.com/user/42"
                        coursesTotal: 12
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
...
  /api/v1/catalog/studios:
    get:
      operationId: listStudios
      summary: List studios with optional search
      description: |
        Returns a paginated list of studios. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. Results are
        ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
      responses:
        "200":
          description: Studio list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvstu0000000000000000001
                        slug: zero-to-mastery
                        displayName: Zero To Mastery
                        externalIds:
                          - source: udemy
                            externalId: "udemy:organization:ztm"
                            url: "https://udemy.com/organization/ztm"
                        coursesTotal: 5
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
...
  /api/v1/catalog/tags:
    get:
      operationId: listTags
      summary: List tags with optional search and category filter
      description: |
        Returns a paginated list of tags. The optional `search` parameter
        performs a case-insensitive substring match on `displayName`. The
        optional `category` parameter filters by exact category value.
        Results are ordered by `displayName` ascending.
      tags: [Entities]
      security:
        - bearerAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Number of items to skip (zero-based).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Maximum number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          required: false
          description: Case-insensitive substring filter on `displayName`.
          schema:
            type: string
        - name: category
          in: query
          required: false
          description: Exact match filter on `category`. Omit to return tags from all categories.
          schema:
            type: string
      responses:
        "200":
          description: Tag list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagListDto"
              examples:
                list:
                  value:
                    items:
                      - id: clxvtag0000000000000000001
                        slug: javascript
                        displayName: JavaScript
                        category: language
                        externalIds: []
                        coursesTotal: 8
                        createdAt: "2026-05-23T10:00:00Z"
                        updatedAt: "2026-05-23T10:00:00Z"
                    total: 1
                    offset: 0
                    limit: 20
        "401":
          description: Missing or invalid bearer token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                unauthorized:
                  value:
                    type: about:blank
                    title: Unauthorized
                    status: 401
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Better Auth bearer token
  schemas:
    BackfillJobAccepted:
      type: object
      required: [jobId]
      description: Confirmation that a backfill job has been accepted and is running in the background.
      properties:
        jobId:
          type: string
          description: Unique cuid identifying the background job. Subscribe to `maintenance:backfill:{jobId}` on Centrifugo for progress events.
      example:
        jobId: clxvjob0000000000000000001
    BackfillMetadataRequest:
      type: object
      description: |
        Optional scope for a metadata backfill job. When both fields are
        omitted the backfill runs across all libraries.
      properties:
        libraryId:
          type: string
          description: cuid of the library to restrict the backfill to. Omit to backfill all libraries.
      example:
        libraryId: clxvlib0000000000000000001
...
    EntitySlug:
      type: string
      pattern: "^[a-z0-9](?:[a-z0-9-]{0,98}[a-z0-9])?$"
      minLength: 1
      maxLength: 100
      description: >-
        URL-safe slug. 1–100 chars, lowercase ASCII letters, digits, and hyphens;
        cannot start or end with a hyphen. Shared by Instructor, Studio, and Tag aggregates.
      example: andrei-neagoie
...
    InstructorDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of an Instructor aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable instructor name.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this instructor.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this instructor.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvins0000000000000000001
        slug: andrei-neagoie
        displayName: Andrei Neagoie
        externalIds:
          - source: udemy
            externalId: "udemy:instructor:42"
            url: "https://udemy.com/user/42"
        coursesTotal: 12
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
...
    StudioDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of a Studio aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable studio name.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this studio.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this studio.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvstu0000000000000000001
        slug: zero-to-mastery
        displayName: Zero To Mastery
        externalIds:
          - source: udemy
            externalId: "udemy:organization:ztm"
            url: "https://udemy.com/organization/ztm"
        coursesTotal: 5
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
...
    TagDto:
      type: object
      required: [id, slug, displayName, externalIds, coursesTotal, createdAt, updatedAt]
      description: Full representation of a Tag aggregate.
      properties:
        id:
          type: string
          description: Server-generated cuid.
        slug:
          $ref: "#/components/schemas/EntitySlug"
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable tag name.
        category:
          oneOf:
            - type: string
              maxLength: 64
            - type: "null"
          description: Free-form category label (e.g. `language`, `framework`, `topic`). Null when uncategorised.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this tag.
        coursesTotal:
          type: integer
          minimum: 0
          description: Total number of courses linked to this tag.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      example:
        id: clxvtag0000000000000000001
        slug: javascript
        displayName: JavaScript
        category: language
        externalIds: []
        coursesTotal: 8
        createdAt: "2026-05-23T10:00:00Z"
        updatedAt: "2026-05-23T10:00:00Z"
...
    UpsertInstructorRequest:
      type: object
      required: [displayName]
      description: Payload for creating or updating an instructor.
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable name displayed in the UI.
        slug:
          $ref: "#/components/schemas/EntitySlug"
          description: URL-safe slug. Auto-generated from `displayName` when omitted.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this instructor.
      example:
        displayName: Andrei Neagoie
        slug: andrei-neagoie
        externalIds:
          - source: udemy
            externalId: "udemy:instructor:42"
            url: "https://udemy.com/user/42"
    UpsertStudioRequest:
      type: object
      required: [displayName]
      description: Payload for creating or updating a studio.
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable studio name.
        slug:
          $ref: "#/components/schemas/EntitySlug"
          description: URL-safe slug. Auto-generated from `displayName` when omitted.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this studio.
      example:
        displayName: Zero To Mastery
        slug: zero-to-mastery
        externalIds:
          - source: udemy
            externalId: "udemy:organization:ztm"
            url: "https://udemy.com/organization/ztm"
    UpsertTagRequest:
      type: object
      required: [displayName]
      description: Payload for creating or updating a tag.
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
          description: Human-readable tag name.
        slug:
          $ref: "#/components/schemas/EntitySlug"
          description: URL-safe slug. Auto-generated from `displayName` when omitted.
        category:
          oneOf:
            - type: string
              maxLength: 64
            - type: "null"
          description: Free-form category label (e.g. `language`, `framework`, `topic`). Pass `null` to clear.
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdRef"
          description: External system references for this tag.
      example:
        displayName: JavaScript
        slug: javascript
        category: language
...
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type:
          type: ["string", "null"]
          format: uri
          default: about:blank
        title:
          type: ["string", "null"]
        status:
          type: integer
          format: int32
        detail:
          type: ["string", "null"]
        instance:
          type: ["string", "null"]
        code:
          type: ["string", "null"]
...